Refactor to use googleApiProvider to fetch data. Refactor the googleApiProviderContext to use the functions from fetcher. Makes cleaner code in my opinion.
